+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/shipment/ShippingAddressBeanX.java
@@ -30,7 +30,7 @@ public class ShippingAddressBeanX {
      */
 
     @JsonProperty("address_id")
-    private int addressId;
+    private long addressId;
     @JsonProperty("address_line")
     private String addressLine;
     @JsonProperty("street_name")
@@ -68,7 +68,7 @@ public class ShippingAddressBeanX {
     @JsonProperty("types")
     private List<String> types;
 
-    public int getAddressId() {
+    public long getAddressId() {
         return addressId;
     }
